Guards implemented and services updated

+++ api/src/features/users/users.controller.ts
@@ -10,17 +10,23 @@ import {
   HttpCode,
   HttpStatus,
   BadRequestException,
+  ForbiddenException,
   NotFoundException,
   UseFilters,
+  UseGuards,
+  Req,
 } from '@nestjs/common'
 
 // Swagger imports
 import {
   ApiTags,
+  ApiBearerAuth,
   ApiOkResponse,
   ApiCreatedResponse,
   ApiNoContentResponse,
   ApiBadRequestResponse,
+  ApiUnauthorizedResponse,
+  ApiForbiddenResponse,
   ApiNotFoundResponse,
   ApiConflictResponse,
   ApiUnprocessableEntityResponse,
@@ -34,6 +40,7 @@ import { UpdateUserDto } from './dto/update-user.dto'
 // Validations / Guards / filters imports
 import { isValidObjectId } from 'mongoose'
 import { MongoExceptionFilter } from '@app/libs/filters'
+import { AccessTokenGuard } from '../auth/guards'
 
 @ApiTags('users')
 @Controller('users')
@@ -72,11 +79,23 @@ export class UsersController {
 
   // Updates a user and returns it
   @Patch(':id')
+  @ApiBearerAuth()
+  @UseGuards(AccessTokenGuard)
   @UseFilters(MongoExceptionFilter)
   @ApiOkResponse({ description: 'OK' })
   @ApiBadRequestResponse({ description: 'Bad Request' })
+  @ApiUnauthorizedResponse({ description: 'Unauthorized' })
+  @ApiForbiddenResponse({ description: 'Forbidden' })
   @ApiConflictResponse({ description: 'Conflict' })
-  async update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  async update(
+    @Req() req,
+    @Param('id') id: string,
+    @Body() updateUserDto: UpdateUserDto,
+  ) {
+    // Checking if the token's sub exists
+    if (!(await this.usersService.findOne(req.user['sub'])))
+      throw new ForbiddenException()
+    // Checking if the user to update id is correct
     if (!isValidObjectId(id))
       throw new BadRequestException('Incorrect id format')
     const updatedUser = await this.usersService.update(id, updateUserDto)
@@ -86,10 +105,18 @@ export class UsersController {
 
   // Deletes a user
   @Delete(':id')
+  @ApiBearerAuth()
+  @UseGuards(AccessTokenGuard)
   @ApiNoContentResponse({ description: 'No Content' })
   @ApiBadRequestResponse({ description: 'Bad Request' })
+  @ApiUnauthorizedResponse({ description: 'Unauthorized' })
+  @ApiForbiddenResponse({ description: 'Forbidden' })
   @HttpCode(HttpStatus.NO_CONTENT)
-  async remove(@Param('id') id: string) {
+  async remove(@Req() req, @Param('id') id: string) {
+    // Checking if the token's sub exists
+    if (!(await this.usersService.findOne(req.user['sub'])))
+      throw new ForbiddenException()
+    // Checking if the user to update id is correct
     if (!isValidObjectId(id))
       throw new BadRequestException('Incorrect id format')
     const deleteResult: any = await this.usersService.remove(id)
